build_image: fall back to ro-mount automatically
A rootfs built with verification cannot be mounted rw, so have the
mount script try to mount things ro if the rw mount failed.

BUG=None
TEST=`./build_image && ./mount.sh` work

Reviewed-on: https://gerrit.chromium.org/gerrit/35065
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>

+++ build_library/disk_layout_util.sh
@@ -158,8 +158,12 @@ EOF
     if [[ ${size} -gt 1 ]]; then
       cat <<-EOF >>"${mount}"
 mkdir -p ${dir}
-sudo mount -o loop,offset=${start_b},sizelimit=${size_b} ${target} ${dir} || \
-  rmdir ${dir}
+m=( sudo mount -o loop,offset=${start_b},sizelimit=${size_b} ${target} ${dir} )
+if ! "\${m[@]}"; then
+  if ! "\${m[@]}" -o ro; then
+    rmdir ${dir}
+  fi
+fi
 EOF
       cat <<-EOF >>"${umount}"
 if [[ -d ${dir} ]]; then
